GuideBot/requirements.txt
chainlit
google.generativeai